+++ src/lib.rs
@@ -1,0 +1,5 @@
+// Compile without standard library
+#![no_std]
+#![no_main]
+
+pub mod board;
